app/recipe/serializers.py
"""
Serializers for recipe API
"""
from rest_framework import serializers

from .models import Recipe, Ingredient


class IngredientSerializer(serializers.ModelSerializer):
    # initialize model and fields to serialize
    class Meta:
        model = Ingredient
        fields = ['name']


class RecipeSerializer(serializers.ModelSerializer):
    ingredients = IngredientSerializer(many=True, required=False)

    class Meta:
        model = Recipe
        fields = '__all__'
        #fields = ['name', 'description', 'ingredients']
